Add derivation of thunk
Thunks have always confused me, and this is solely due to the fact that I never fully understood the word. Just knowing the specialized sense of a word does not clear up confusion when you have no idea where the word came from.

I have always been asking "but *why* is it called thunk?" This has honestly blown me off of Redux entirely for two years. I finally took the effort to find out.

I'm proposing we add the derivation of the term and its source, which will help greatly in clearing up understanding for people.