Move Redux module type extension into a separate imported file
We previously merged a PR that would globally override the types
from the Redux core package and patch the `Dispatch` type to
_always_ assume that the thunk middleware is installed.

However, that assumption is not always true, and globally
overriding types is a dangerous approach to have as a default.

This commit moves the global overrides into a separate types file,
which can now be imported as 'redux-thunk/extend-redux'.  This
allows users to still opt in to that behavior if desired.

The typetest files have been updated to compare the behavior with
and without importing the override file.